Module 2

+++ src/controllers/cards.ts
@@ -1,60 +1,63 @@
-import { Request, Response } from "express";
-import { request } from "http";
-import mongoose from "mongoose";
-import { Card } from "../models/card";
-import { RequestCustom } from "../types/types";
-const { PORT = 123, DB_URL = "123" } = process.env;
 
+import { NextFunction, Request, Response } from 'express';
+import mongoose from 'mongoose';
+import Card from '../models/card';
+import { RequestCustom } from '../types/types';
+import BadRequestErr from '../errors/bad-Request';
+import NotFoundErr from '../errors/not-found';
+import AccessErr from '../errors/access';
 
+export const GetCards = async (req: RequestCustom, res: Response, next:NextFunction) => {
 
-
-
-export const GetCards = async (req: RequestCustom, res: Response) => {
   try {
     const cards = await Card.find({});
     return res.status(200).send(cards);
   } catch (err) {
 
     console.log(err);
-    return res.status(500).send({ message: "ошибка сервера" });
+    next(err);
   }
 };
-export const CreateCard = async (req: RequestCustom, res: Response) => {
+
+export const CreateCard = async (req: RequestCustom, res: Response, next:NextFunction) => {
+
   try {
     const newCard = await Card.create({
       name: req.body.name,
       link: req.body.link,
       createdAt: new Date(),
       likes: [],
-      owner: req.body.user?._id,
+      owner: req.user?._id,
     });
     return res.status(200).send(newCard);
   } catch (err) {
     if (err instanceof mongoose.Error.ValidationError) {
-      return res.status(400).send({ message: err.message });
+
+      next(new BadRequestErr(err.message));
+    } else {
+      console.log(err);
+      next(err);
     }
-    console.log(err);
-    return res.status(500).send({ message: "ошибка сервера" });
   }
 };
-export const DeleteCard = async (req: Request, res: Response) => {
+export const DeleteCard = async (req: RequestCustom, res: Response, next:NextFunction) => {
   try {
     const { cardId } = req.params;
+    const cardOwner = await Card.findById(cardId);
+    if (req.user?._id.toString() !== cardOwner?.owner.toString()) {
+      throw new AccessErr('Ошибка доступа, это не ваша карточка');
+    }
     const card = await Card.findByIdAndRemove(cardId);
     if (!card) {
-      // const error = new Error("Такого пользователя не существует");
-      // error.name = "Not found";
-      // throw error;
-      throw res.status(404).send({ message: "Такого карточки не существует" })
-
-     }
+      throw new NotFoundErr('Такой карточки не существует');
+    }
     return res.status(200).send(card);
   } catch (err) {
     console.log(err);
-    return res.status(500).send({ message: "ошибка сервера" });
+    next(err);
   }
 };
-export const GetLike = async (req: Request, res: Response) => {
+export const GetLike = async (req: Request, res: Response, next:NextFunction) => {
   try {
     const { cardId } = req.params;
     const card = await Card.findByIdAndUpdate(
@@ -63,48 +66,38 @@ export const GetLike = async (req: Request, res: Response) => {
       { new: true },
     );
     if (!card) {
-      // const error = new Error("Такого пользователя не существует");
-      // error.name = "Not found";
-      // throw error;
-      throw res.status(404).send({ message: "Такого карточки не существует" })
-
-     }
+      throw new NotFoundErr('Такой карточки не существует');
+    }
     return res.status(200).send(card);
   } catch (err) {
     if (err instanceof mongoose.Error.ValidationError) {
-      return res.status(400).send({ message: err.message });
+      next(new BadRequestErr(err.message));
+    } else {
+      console.log(err);
+      next(err);
     }
-    console.log(err);
-    return res.status(500).send({ message: "ошибка сервера" });
   }
 };
+export const DeleteLike = async (req: Request, res: Response, next:NextFunction) => {
 
-
-
-
-
-/////исправить удаление лайка
-export const DeleteLike = async (req: Request, res: Response) => {
   try {
     const { cardId } = req.params;
     const card = await Card.findByIdAndUpdate(
       cardId,
       { $pull: { likes: req.body.user._id } },
       { new: true },
-    )
+    );
     if (!card) {
-      // const error = new Error("Такого пользователя не существует");
-      // error.name = "Not found";
-      // throw error;
-      throw res.status(404).send({ message: "Такого карточки не существует" })
-
-     }
+      throw new NotFoundErr('Такой карточки не существует');
+    }
     return res.status(200).send(card);
   } catch (err) {
     if (err instanceof mongoose.Error.ValidationError) {
-      return res.status(400).send({ message: err.message });
+      next(new BadRequestErr(err.message));
+    } else {
+      console.log(err);
+      next(err);
     }
-    console.log(err);
-    return res.status(500).send({ message: "ошибка сервера" });
+
   }
 };
